+++ sentry-rails/lib/sentry/rails/configuration.rb
@@ -20,9 +20,9 @@ module Sentry
       'ActionController::RoutingError',
       'ActionController::UnknownAction',
       'ActionController::UnknownFormat',
+      'ActionDispatch::Http::MimeNegotiation::InvalidType',
       'ActionController::UnknownHttpMethod',
       'ActionDispatch::Http::Parameters::ParseError',
-      'ActiveJob::DeserializationError', # Can cause infinite loops
       'ActiveRecord::RecordNotFound'
     ].freeze
     class Configuration