+++ sentry-rails/spec/sentry/rails/activejob_spec.rb
@@ -1,14 +1,21 @@
 require "spec_helper"
+require "active_job/railtie"
 
-class MyActiveJob < ActiveJob::Base
+class FailedJob < ActiveJob::Base
   self.logger = nil
 
   class TestError < RuntimeError
   end
 
   def perform
-    Sentry.get_current_scope.set_extras(foo: :bar)
     raise TestError, "Boom!"
+  end
+end
+
+class MyActiveJob < FailedJob
+  def perform
+    Sentry.get_current_scope.set_extras(foo: :bar)
+    super
   end
 end
 
@@ -18,9 +25,21 @@ class RescuedActiveJob < MyActiveJob
   def rescue_callback(error); end
 end
 
+RSpec.describe "without Sentry initialized" do
+  before(:each) do
+    FailedJob.queue_adapter = :inline
+  end
+
+  it "runs job" do
+    job = FailedJob.new
+
+    expect { job.perform_now }.to raise_error(FailedJob::TestError)
+  end
+end
+
 RSpec.describe "ActiveJob integration" do
   before(:all) do
-    perform_basic_setup
+    make_basic_app
   end
 
   let(:event) do
@@ -49,6 +68,75 @@ RSpec.describe "ActiveJob integration" do
     expect(event["extra"]["foo"]).to eq("bar")
 
     expect(Sentry.get_current_scope.extra).to eq({})
+  end
+
+  context "when DeserializationError happens in user's jobs" do
+    before do
+      make_basic_app
+    end
+
+    class DeserializationErrorJob < ActiveJob::Base
+      def perform
+        1/0
+      rescue
+        raise ActiveJob::DeserializationError
+      end
+    end
+
+    it "reports the root cause to Sentry" do
+      expect do
+        DeserializationErrorJob.perform_now
+      end.to raise_error(ActiveJob::DeserializationError, /divided by 0/)
+
+      expect(transport.events.size).to eq(1)
+
+      event = transport.events.last.to_json_compatible
+      expect(event.dig("exception", "values", 0, "type")).to eq("ZeroDivisionError")
+    end
+
+    context "and in user-defined reporting job too" do
+      before do
+        Sentry.configuration.async = lambda do |event, hint|
+          UserDefinedReportingJob.perform_now(event, hint)
+        end
+      end
+
+      class UserDefinedReportingJob < ActiveJob::Base
+        def perform(event, hint)
+          Post.find(0)
+        rescue
+          raise ActiveJob::DeserializationError
+        end
+      end
+
+      it "doesn't cause infinite loop because of excluded exceptions" do
+        expect do
+          DeserializationErrorJob.perform_now
+        end.to raise_error(ActiveJob::DeserializationError, /divided by 0/)
+      end
+    end
+
+    context "and in customized SentryJob too" do
+      before do
+        class CustomSentryJob < ::Sentry::SendEventJob
+          def perform(event)
+            raise "Not excluded exception"
+          rescue
+            raise ActiveJob::DeserializationError
+          end
+        end
+
+        Sentry.configuration.async = lambda do |event, hint|
+          CustomSentryJob.perform_now(event, hint)
+        end
+      end
+
+      it "doesn't cause infinite loop" do
+        expect do
+          DeserializationErrorJob.perform_now
+        end.to raise_error(ActiveJob::DeserializationError, /divided by 0/)
+      end
+    end
   end
 
   context 'using rescue_from' do